requirements.txt
aioredis==2.0.1
async-timeout==4.0.2
black==23.1.0
click==8.1.3
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
python-environ==0.4.54
tomli==2.0.1
typing_extensions==4.5.0
